+++ api/admin/internal/logic/product/feighttemplate/feighttemplateupdatelogic.go
@@ -31,10 +31,10 @@ func (l *FeightTemplateUpdateLogic) FeightTemplateUpdate(req types.UpdateFeightT
 		Id:             req.Id,
 		Name:           req.Name,
 		ChargeType:     req.ChargeType,
-		FirstWeight:    int64(req.FirstWeight),
-		FirstFee:       int64(req.FirstFee),
-		ContinueWeight: int64(req.ContinueWeight),
-		ContinmeFee:    int64(req.FirstFee),
+		FirstWeight:    req.FirstWeight,
+		FirstFee:       req.FirstFee,
+		ContinueWeight: req.ContinueWeight,
+		ContinmeFee:    req.FirstFee,
 		Dest:           req.Dest,
 	})
